udpdated stock management gui
